Add ImageMountain FishEyeImageVR and TinBuilder Modified PointCloudLODGenerator to use pixelLOD instead of rangeLOD

+++ OsgMFC/OsgFrameWork/MDIClient.cpp
@@ -5,18 +5,18 @@
 #include "OsgFrameWork.h"
 #include "MDIClient.h"
 
-
 // CMDIClient
 
 IMPLEMENT_DYNAMIC(CMDIClient, CWnd)
 
 CMDIClient::CMDIClient()
 {
-
+	bmp.LoadBitmap(IDB_BITMAP_BG);
 }
 
 CMDIClient::~CMDIClient()
 {
+	bmp.Detach();
 }
 
 
@@ -33,38 +33,50 @@ END_MESSAGE_MAP()
 void CMDIClient::OnPaint()
 {
 	CPaintDC dc(this); // device context for painting
-	// TODO: 在此处添加消息处理程序代码
-	// 不为绘图消息调用 CWnd::OnPaint()
 
-	CDC		*pDC = GetDC();   //必须自己使用GetDC()，不要用CPaintDC。  
-	CRect	ClientRect;  
-	GetClientRect(ClientRect);  
-/*
-	pDC->FillSolidRect(ClientRect, GetSysColor(COLOR_APPWORKSPACE));
-	//下面用文字示例，显示位图一样。  
-	CFont	font;
-	font.CreatePointFont(500,   "黑体");  
-	CFont	*fontOld   =   pDC->SelectObject(&font);  
-	pDC->SetBkMode(TRANSPARENT);  
-	pDC->SetTextAlign(TA_BOTTOM);  
-	pDC->SetTextColor(RGB(200,   200,   200));  
-	pDC->TextOut(ClientRect.left   +   1,   ClientRect.bottom,   "客户区LOGO",   10);  
-	pDC->SetTextColor(RGB(100,   100,   100));  
-	pDC->TextOut(ClientRect.left,   ClientRect.bottom   -   1,   "客户区LOGO",   10);  
-	pDC->SelectObject(fontOld);  
-	font.DeleteObject();  
-	ReleaseDC(pDC);   */
-	CBitmap   bmp;  
-	if(bmp.LoadBitmap(IDB_BITMAP_BG))  
-	{  
-		BITMAP	bmpInfo;  
-		bmp.GetBitmap(&bmpInfo);  
-		CDC	dcMemory;  
+	CDC		*pDC = GetDC();   //必须自己使用GetDC()，不要用CPaintDC
+	CRect	ClientRect;
+	GetClientRect(ClientRect);
+
+	
+	{
+		BITMAP	bmpInfo;
+		bmp.GetBitmap(&bmpInfo);
+		CDC	dcMemory;
 		dcMemory.CreateCompatibleDC(pDC);
 		dcMemory.SelectObject(&bmp);
 		pDC->SetStretchBltMode(BLACKONWHITE);
 		pDC->StretchBlt(ClientRect.left,ClientRect.top,ClientRect.Width(),ClientRect.Height(),
 			&dcMemory,0,0,bmpInfo.bmWidth,bmpInfo.bmHeight,SRCCOPY);
-		bmp.Detach();
+	}
+
+	dirty=true;
+	UpdateConsoleOuts();
+}
+
+void CMDIClient::UpdateConsoleOuts()
+{
+	if(dirty) {
+		dirty = false;
+
+		CDC		*pDC = GetDC();
+		CRect	ClientRect;  
+		GetClientRect(ClientRect);  
+		CFont	font;
+		int size = 150;
+		font.CreatePointFont(size,   "Consolas");
+		CFont	*fontOld   =   pDC->SelectObject(&font);  
+		pDC->SetBkMode(TRANSPARENT); 
+		pDC->SetTextAlign(TA_BOTTOM);
+		pDC->SetTextColor(RGB(0, 255, 0));
+		
+		for(int i=0; i<consoleOuts.size(); ++i) {
+			pDC->TextOut(ClientRect.left + 5,
+				ClientRect.top + 18*(i+1)+5,   CString(consoleOuts.at(i).c_str()));
+		}
+
+		pDC->SelectObject(fontOld);
+		font.DeleteObject();
+		ReleaseDC(pDC);
 	}
 }
